Removes modifier file for interface methods

+++ src/main/java/net/h34t/temporize/Temporize.java
@@ -17,7 +17,7 @@ public class Temporize {
 
         try {
             if (args.length < 2) {
-                System.out.println("Usage: java -jar temporize.jar tpl/ output/ path/to/Modifiers.java");
+                System.out.println("Usage: java -jar temporize.jar tpl/ output/");
                 System.out.println("       Compiles all templates with a file name according to the pattern");
                 System.out.println("       [name].temporize.[ending] into precompiled templates and saves");
                 System.out.println("       the java source files to output.");
@@ -27,9 +27,8 @@ public class Temporize {
 
                 File inDirectory = new File(args[0]);
                 File outDirectory = new File(args[1]);
-                String modifier = args.length > 2 ? args[2] : null;
 
-                verbose = Arrays.stream(args).anyMatch(a -> a.equals("--verbose"));
+                verbose = Arrays.asList(args).contains("--verbose");
 
                 if (!inDirectory.exists()) {
                     System.err.println("Couldn't find input directory " + inDirectory.getName());
@@ -48,14 +47,17 @@ public class Temporize {
                 }
 
                 // copy the TemporizeTemplate file from the packaged resources into the output directory
+                // but only if it doesn't exist yet
                 File temporizeInterface = new File(temporizeDirectory, "TemporizeTemplate.java");
 
-                try (InputStream is = Temporize.class.getClassLoader().getResourceAsStream("TemporizeTemplate.java");
-                     OutputStream os = new FileOutputStream(temporizeInterface)) {
-                    byte[] buffer = new byte[1024];
-                    int bytesRead;
-                    while ((bytesRead = is.read(buffer)) != -1) {
-                        os.write(buffer, 0, bytesRead);
+                if (!temporizeInterface.exists()) {
+                    try (InputStream is = Temporize.class.getClassLoader().getResourceAsStream("TemporizeTemplate.java");
+                         OutputStream os = new FileOutputStream(temporizeInterface)) {
+                        byte[] buffer = new byte[1024];
+                        int bytesRead;
+                        while ((bytesRead = is.read(buffer)) != -1) {
+                            os.write(buffer, 0, bytesRead);
+                        }
                     }
                 }
 
@@ -76,9 +78,8 @@ public class Temporize {
 
                                 List<Token> tokens = Parser.FULL.parse(t);
                                 ASTNode root = new ASTBuilder().build(tokens);
-                                Template tpl = new Compiler().compile(packageName, className, modifier, root, inc -> {
-                                    System.out.println(" * Includes " + inc);
-                                });
+                                Template tpl = new Compiler().compile(packageName, className, root,
+                                        inc -> System.out.println(" * Includes " + inc));
 
                                 File packageOutDirectory = tf.getOutputDirectory(outDirectory);
                                 if (!packageOutDirectory.exists() && !packageOutDirectory.mkdirs()) {
